Fix iarray core warnings

+++ src/iarray_constructor.c
@@ -817,7 +817,7 @@ INA_API(ina_rc_t) iarray_copy(iarray_context_t *ctx,
         (*dest)->auxshape = (iarray_auxshape_t *) ina_mem_alloc(sizeof(iarray_auxshape_t));
         for (int i = 0; i < (*dest)->dtshape->ndim; ++i) {
             (*dest)->auxshape->offset[i] = 0;
-            (*dest)->auxshape->index[i] = i;
+            (*dest)->auxshape->index[i] = (int8_t) i;
             (*dest)->auxshape->shape_wos[i] = src->dtshape->shape[i];
             (*dest)->auxshape->pshape_wos[i] = src->dtshape->pshape[i];
         }
